+++ packages/yourloops/test/unit/pages/profile/profile-page-context.hook.test.ts
@@ -28,7 +28,6 @@
 import { act, renderHook } from '@testing-library/react'
 import * as authHookMock from '../../../../lib/auth'
 import { type User } from '../../../../lib/auth'
-import * as patientContext from '../../../../lib/patient/patient.provider'
 import * as alertMock from '../../../../components/utils/snackbar'
 import useProfilePageContextHook from '../../../../pages/profile/profile-page-context.hook'
 import { type Profile } from '../../../../lib/auth/models/profile.model'
@@ -41,7 +40,6 @@ import { LanguageCodes } from '../../../../lib/auth/models/enums/language-codes.
 import { Unit } from 'medical-domain'
 
 jest.mock('../../../../lib/auth')
-jest.mock('../../../../lib/patient/patient.provider')
 jest.mock('../../../../components/utils/snackbar')
 
 describe('Profile page context hook', () => {
@@ -78,12 +76,8 @@ describe('Profile page context hook', () => {
   const updateProfileMock = jest.fn()
   const updateSettingsMock = jest.fn()
   const updatePreferencesMock = jest.fn()
-  const refreshPatientMock = jest.fn()
 
   beforeEach(() => {
-    (patientContext.usePatientContext as jest.Mock).mockImplementation(() => ({
-      refresh: refreshPatientMock
-    }));
     (alertMock.useAlert as jest.Mock).mockImplementation(() => ({
       success: onSuccessAlertMock,
       error: onErrorAlertMock
@@ -129,7 +123,6 @@ describe('Profile page context hook', () => {
     expect(updateProfileMock).toHaveBeenCalledWith(expectedProfile)
     expect(updateSettingsMock).toHaveBeenCalledWith(expectedSettings)
     expect(updatePreferencesMock).toHaveBeenCalledWith(expectedPreferences)
-    expect(refreshPatientMock).toHaveBeenCalled()
     expect(onSuccessAlertMock).toHaveBeenCalled()
   })